Move is_script(), is_report() into extras.utils
